\name{G519}
\alias{G519}
\title{Granta-519 Data}
\description{
  LRR for Granta-519 Cell Line data (published in DeLeeuw et al. (2004). Comprehensive whole genome array CGH profiling of mantle cell lymphoma model genomes, Human Molecular Genetics, 13(17):1827-1837).
}
\usage{data(G519)}
\format{A matrix containing four columns: chromosome, start and end probe positions, observed Log R Ratio.
}
